refactor waitUntilWorkDone into TestConquery
