+++ backend/src/test/java/com/bakdata/conquery/integration/common/LoadingUtil.java
@@ -65,7 +65,7 @@ public class LoadingUtil {
 
 		// wait only if we actually did anything
 		if (!content.getPreviousQueryResults().isEmpty()) {
-			support.waitUntilWorkDone();
+			support.testConquery.waitUntilWorkDone();
 		}
 	}
