+++ backend/src/test/java/com/bakdata/conquery/integration/tests/FilterResolutionContainsTest.java
@@ -55,7 +55,7 @@ public class FilterResolutionContainsTest extends IntegrationTest.Simple impleme
 		FilterSearch
 			.updateSearch(conquery.getNamespace().getNamespaces(), Collections.singleton(conquery.getNamespace().getDataset()), conquery.getDatasetsProcessor().getJobManager());
 
-		conquery.waitUntilWorkDone();
+		conquery.testConquery.waitUntilWorkDone();
 
 		VirtualConcept concept = (VirtualConcept) conquery.getNamespace().getStorage().getAllConcepts().iterator().next();
 		VirtualConceptConnector connector = concept.getConnectors().iterator().next();
